Fix bug with invalid credentials
	* refctoring
