+++ bin/console
@@ -1,6 +1,5 @@
 #!/usr/bin/env ruby
 
-require 'bundler/setup'
 require 'admitad'
 require 'yaml'
 require 'pry'
