+++ lib/admitad/version.rb
@@ -1,3 +1,3 @@
 module Admitad
-  VERSION = '0.0.11'.freeze
+  VERSION = '0.0.12'.freeze
 end
